Fixed the UInt width problems in TileLink.scala.
These problems are made when migrating from chisel2 to chisel3.
In chisel2, we use UInt(value, width) to define a fixed width UInt.
In chisel3, it should be value.U(width.W).
But in TileLink.scala, I turned them into value.U(width),
which means take the width bit of value.U!!!

So, be careful when you are copy-pasting code from rocketchip or boom.