mem,sq: don't set commited flag when mmio commits
For simplicity, we increase cmtPtr and set commited flags when
store instructions commit from ROB. However, this causes problems
when there's an MMIO at the head and new instructions enqueue after
the MMIO finishes but before the MMIO commits.

For example, at cycle T0 mmio store instruction with sqPtr p0
finishes and deqPtr is increased. At cycle T1, a new store
instruction enqueues and the p0 is allocated for this new instruction.
At cycle T2, the mmio instruction commits from ROB. With cmtPtr,
p0 is set to commited. However, since p0 is allocated to the new
store instruction, the commited flag should be false.

The example causes at least two issues. (1) when redirect happens, p0
will not be flushed. (2) p0 will be commit to sbuffer before p0
writes back to store queue.

In this commit, we change how commited flag works. We add a
`hasInflightMMIO` flag. When store instructions commit, we only
set commited flag when `hasInflightMMIO` is not set.