rob: fix walkCounter when all entries are flushed (#1125)
This commit fixes the walkCounter value in Rob when every instruction is
flushed.

For example, when rob is full and the head instruction causes an
exception, the redirect robIdx is the deqPtr. In this case,
currentWalkPtr is enqPtr - 1.U and redirectWalkDistance is RobSize - 1.
Since exceptions flush the instruction itself, flushItSelf is true.B.
Previously we use `+` to count the walk distance and it causes overflows
when RobSize is power of 2. We change it to `+&` to allow walkCounter to
be RobSize. The width of walkCounter also needs to be changed.