rename: don't update refCounter 0 (#1126)
This commit removes the update logic for ref counter 0.

For simplicity, we don't count the number of references for physical
register 0. It should never be released to freelist.

Previously we track register 0's references. It works fine but it makes
the performance counters confusing because it may increase to a large
number. It never causes real issues.