rob: fix commitInstr performance counters (#1679)
This commit fixes the wrong update values of commitInstr in ROB.
Previously we add a RegNext to trueCommitCnt in #1644. However,
we don't add RegNext to the ifCommit condition. This makes the
commitInstr larger than the normal value and IPC looks better.

This commit fixes this bug and the IPC numbers should be correct now.