config: use smaller BPU in MinimalConfig (#3029)
EnableBPD is not used by BPU anymore. Remove it in Config.
This PR adds the override config of BPU to MinimalConfig.

This configuration aims to provide very different parameters from
DefaultConfig. Thus allowing verification of more parameters.
The overall size of BPU is also reduced, which slightly reduces the
compile time (around 20s on 7950X).