StoreQueue: fix X  when write StoreBuffer (#3068)
If there is an inactive element in a vector access instruction, it will
not be sent to the sta pipeline, so the data in the corresponding entry
in the StoreQueue is invalid, and we use `vecValid` to differentiate
between valid and invalid data in table entries.

For StoreQueue entries, `vecValid` depends on `hasException`,
`vecDataValid` (whether the data in the table entry is valid or not),
and `isVec` (whether it is a table entry for a vector instruction or
not). `vecDataValid` is initialized to false, and will be set to true
when sta writes back to the entry and the entry is a vector instruction.

Only vector instructions have invalid data in table entries, because we
can't determine how many table entries are needed for a vector access
uop at dispatch time.