LSU: trigger accessfault for vector mmio access request (#3168)
This commit is to ensure robustness when vector instructions are used to
access MMIO regions by mistake. Currently, we do not support vector
instruction access to MMIO regions, and in order to handle the situation
when a vector instruction incorrectly accesses an MMIO region, an
accessfault exception is triggered when the MMIO region is accessed
using a vector instruction.