refactor(LoadPipe): remove the redundant logic of the `mq_nack` (#3936)
Remove the redundant logic of the miss queue nack based on the issues
raised in the
issues(https://github.com/OpenXiangShan/XiangShan/issues/3916).

When a tlb miss and a dcache miss occur at the same time and the miss
queue nack, it will cause the `LoadUnit` to generate both replay signals
`C_TM` and `C_DR`. We will give priority to `C_TM`, which is why we need
to send a kill signal to dcache when a tlb miss occurs.

Although there was no problem before, as the
issue(https://github.com/OpenXiangShan/XiangShan/issues/3916) says, this
will cause ambiguity, and the miss queue nack message is already
included in `s2_nack_no_mshr`, so the choice is to remove the
`s2_miss_req_fire` signal from the generation logic of the `s2_mq_nack`
signal.