ci: fix date error in performance regression (#4347)
The workflow uses `date +'%V'` to get the week number, but the ISO week
number has a leading zero, which is interpreted as an octal number by
bash. This caused errors in weeks 08 and 09, which went unnoticed
previously.

This patch fixes the issue by using `10#$(date +'%V')` to force the
number to be interpreted as decimal.